Rename category router to 'collection'
