+++ server/routes/api.js
@@ -2,14 +2,15 @@ const express = require('express');
 const router  = express.Router();
 
 const { isLoggedIn } = require('../controllers/authController');
-const authRouter     = require('./auth');
-const categoryRouter = require('./category');
-const feedRouter     = require('./feed');
-const itemRouter = require('./item');
+
+const authRouter       = require('./auth');
+const collectionRouter = require('./collection');
+const feedRouter       = require('./feed');
+const itemRouter       = require('./item');
 
 // api routes
 router.use('/auth', authRouter);
-router.use('/categories', isLoggedIn, categoryRouter);
+router.use('/categories', isLoggedIn, collectionRouter);
 router.use('/feeds', isLoggedIn, feedRouter);
 router.use('/feeditems', isLoggedIn, itemRouter);
